refactor: merge McpEditorWindowEventHandler into McpEditorWindow
