--- Packages/src/Editor/UI/McpEditorWindowEventHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d171c9fe3128441419f4a5b1dbc41ab2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
